+++ src/advanced/features/cart-management/index.ts
@@ -1,0 +1,1 @@
+export { useCartService } from './model/useCartService';
