+++ src/advanced/features/payment/index.ts
@@ -1,0 +1,1 @@
+export { usePaymentService } from './model/usePaymentService';
